+++ 4lab/LoginControl.xaml
@@ -1,0 +1,31 @@
+﻿<UserControl x:Class="_4lab.LoginControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             mc:Ignorable="d" 
+             d:DesignHeight="300" d:DesignWidth="350">
+    <StackPanel>
+        <TextBlock Text="{DynamicResource EmailLabel}" Foreground="Black" FontWeight="Bold"/>
+        <TextBox x:Name="EmailBox"
+                 Background="Black"
+                 Foreground="White"
+                 Margin="0,5,0,15"
+                 Padding="8"
+                 BorderThickness="0"/>
+        <TextBlock Text="{DynamicResource PasswordLabel}" Foreground="Black" FontWeight="Bold"/>
+        <PasswordBox x:Name="PasswordBox"
+                     Background="Black"
+                     Foreground="White"
+                     Margin="0,5,0,15"
+                     Padding="8"
+                     BorderThickness="0"/>
+        <Button x:Name="LoginButton"
+                Content="{DynamicResource LoginButton}"
+                Style="{StaticResource OrangeBlackButtonStyle}"
+                HorizontalAlignment="Center"
+                Margin="0,20,0,10"
+                Padding="20,10"
+                Click="LoginButton_Click"/>
+    </StackPanel>
+</UserControl>
