+++ 4lab/LoginControl.xaml.cs
@@ -1,0 +1,70 @@
+﻿using System.Linq;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Media;
+using _4lab.DB;
+using Roles;
+
+namespace _4lab
+{
+    public partial class LoginControl : UserControl
+    {
+        public LoginControl()
+        {
+            InitializeComponent();
+        }
+
+        private void LoginButton_Click(object sender, RoutedEventArgs e)
+        {
+            string email = EmailBox.Text;
+            string password = PasswordBox.Password;
+
+            if (string.IsNullOrWhiteSpace(email) || string.IsNullOrWhiteSpace(password))
+            {
+                MessageBox.Show("Email и пароль должны быть заполнены.", "Ошибка",
+                    MessageBoxButton.OK, MessageBoxImage.Error);
+                return;
+            }
+
+            using (var context = new _4lab.BD.DBContext())
+            {
+                var user = context.Users.FirstOrDefault(u => u.Email == email);
+                if (user == null || !DataBaseInteractor.VerifyPassword(password, user.PasswordHash))
+                {
+                    MessageBox.Show("Неверный email или пароль.", "Ошибка",
+                        MessageBoxButton.OK, MessageBoxImage.Error);
+                    return;
+                }
+
+                Player player = null;
+                if (user.Role == UserRole.Player)
+                {
+                    player = context.Players.FirstOrDefault(p => p.Id == user.Id);
+                }
+
+                CurrentUser.Instance.Login(null, player);
+                MessageBox.Show("Вход выполнен успешно!", "Успех",
+                    MessageBoxButton.OK, MessageBoxImage.Information);
+                var page = this.FindAncestor<Page>();
+                page?.NavigationService.Navigate(new MainMenuPage((MainWindow)Window.GetWindow(this)));
+            }
+        }
+    }
+
+    // Helper extension method to find ancestor
+    public static class DependencyObjectExtensions
+    {
+        public static T FindAncestor<T>(this DependencyObject dependencyObject) where T : DependencyObject
+        {
+            while (dependencyObject != null)
+            {
+                if (dependencyObject is T target)
+                    return target;
+                dependencyObject = LogicalTreeHelper.GetParent(dependencyObject) ??
+                                 VisualTreeHelper.GetParent(dependencyObject);
+            }
+            return null;
+        }
+    }
+
+}
