+++ 4lab/RegisterControl.xaml
@@ -1,0 +1,45 @@
+﻿<UserControl x:Class="_4lab.RegisterControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             mc:Ignorable="d" 
+             d:DesignHeight="400" d:DesignWidth="350">
+    <StackPanel>
+        <TextBlock Text="{DynamicResource UsernameLabel}" Foreground="Black" FontWeight="Bold"/>
+        <TextBox x:Name="UsernameBox"
+                 Background="Black"
+                 Foreground="White"
+                 Margin="0,5,0,15"
+                 Padding="8"
+                 BorderThickness="0"/>
+        <TextBlock Text="{DynamicResource EmailLabel}" Foreground="Black" FontWeight="Bold"/>
+        <TextBox x:Name="EmailBox"
+                 Background="Black"
+                 Foreground="White"
+                 Margin="0,5,0,15"
+                 Padding="8"
+                 BorderThickness="0"/>
+        <TextBlock Text="{DynamicResource PasswordLabel}" Foreground="Black" FontWeight="Bold"/>
+        <PasswordBox x:Name="PasswordBox"
+                     Background="Black"
+                     Foreground="White"
+                     Margin="0,5,0,15"
+                     Padding="8"
+                     BorderThickness="0"/>
+        <TextBlock Text="{DynamicResource ConfirmPasswordLabel}" Foreground="Black" FontWeight="Bold"/>
+        <PasswordBox x:Name="ConfirmPasswordBox"
+                     Background="Black"
+                     Foreground="White"
+                     Margin="0,5,0,15"
+                     Padding="8"
+                     BorderThickness="0"/>
+        <Button x:Name="RegisterButton"
+                Content="{DynamicResource RegisterButton}"
+                Style="{StaticResource OrangeBlackButtonStyle}"
+                HorizontalAlignment="Center"
+                Margin="0,20,0,0"
+                Padding="20,10"
+                Click="RegisterButton_Click"/>
+    </StackPanel>
+</UserControl>
