+++ 4lab/RegisterControl.xaml.cs
@@ -1,0 +1,86 @@
+﻿using System;
+using System.Linq;
+using System.Text.RegularExpressions;
+using System.Windows;
+using System.Windows.Controls;
+using _4lab.DB;
+using Roles;
+
+namespace _4lab
+{
+    public partial class RegisterControl : UserControl
+    {
+        public RegisterControl()
+        {
+            InitializeComponent();
+        }
+
+        private void RegisterButton_Click(object sender, RoutedEventArgs e)
+        {
+            string username = UsernameBox.Text;
+            string email = EmailBox.Text;
+            string password = PasswordBox.Password;
+            string confirmPassword = ConfirmPasswordBox.Password;
+
+            if (string.IsNullOrWhiteSpace(username) || string.IsNullOrWhiteSpace(email) ||
+                string.IsNullOrWhiteSpace(password) || string.IsNullOrWhiteSpace(confirmPassword))
+            {
+                MessageBox.Show("Все поля должны быть заполнены.", "Ошибка",
+                    MessageBoxButton.OK, MessageBoxImage.Error);
+                return;
+            }
+
+            if (!Regex.IsMatch(email, @"^[^@\s]+@[^@\s]+\.[^@\s]+$"))
+            {
+                MessageBox.Show("Введите корректный email.", "Ошибка",
+                    MessageBoxButton.OK, MessageBoxImage.Error);
+                return;
+            }
+
+            if (password != confirmPassword)
+            {
+                MessageBox.Show("Пароли не совпадают.", "Ошибка",
+                    MessageBoxButton.OK, MessageBoxImage.Error);
+                return;
+            }
+
+            string hashedPassword = DataBaseInteractor.HashPassword(password);
+
+            try
+            {
+                using (var context = new _4lab.BD.DBContext())
+                {
+                    if (context.Users.Any(u => u.Email == email))
+                    {
+                        MessageBox.Show("Пользователь с таким email уже существует.", "Ошибка",
+                            MessageBoxButton.OK, MessageBoxImage.Error);
+                        return;
+                    }
+
+                    var user = new Player
+                    {
+                        Username = username,
+                        Email = email,
+                        PasswordHash = hashedPassword,
+                        Role = UserRole.Player,
+                    };
+
+                    context.Users.Add(user);
+                    context.SaveChanges();
+
+                    CurrentUser.Instance.SetUser(user);
+                    MessageBox.Show("Регистрация прошла успешно!", "Успех",
+                        MessageBoxButton.OK, MessageBoxImage.Information);
+
+                    var page = this.FindAncestor<Page>();
+                    page?.NavigationService.Navigate(new MainMenuPage((MainWindow)Window.GetWindow(this)));
+                }
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show($"Ошибка при регистрации: {ex.Message}", "Ошибка",
+                    MessageBoxButton.OK, MessageBoxImage.Error);
+            }
+        }
+    }
+}
